added a new function 'prep_data()' to format and arrange initial data
'prep_data()', the new function will take what was step 0 (preparing data) and the development of permutations, so that it returns the results matrix with pseudo-F and the mean squares. This data frame can be used either by 'scompvar()' or 'sim_beta()', so they can be used simultaneously instead of following a given sequence.