Update functions prep_data and plot_power
`prep_data()` is now a function that handles the pre-processing for data in both models (i.e. single.factor and nested.symmetric). This has become the first function to use as soon as the package is to be used.
`plot_power()` is improved by showing a comparison of the different power progressions as m and n are changed.