sim_beta now differentiates between single factor and nested
sim_beta is now able to work with both frameworks adequately, it results in data that is listed with n for single factor model or n + m in nested model.
Minor adjustments to prep_data that eliminate the column m for single factor model.
Minor adjustments for scompvar that improve the management differences between single factor and nested model