plot_power is divided in two files
The function `plot_power()` was becoming too large, so a second file was devised to only include the helper functions, thus streamlining the original `plot_power()`

plot_power0: The new file includes the definition for the functions to prepare and plot the power curves and the density plots. These functions are now ready to work differently for the single factor and nested models.